extracted ethsim to a testkit
