+++ projects/libs/core-library-typescript/src/test-kit/index.ts
@@ -2,3 +2,4 @@ export { FakeGossipClient } from "./fake-gossip-client";
 export { generateServiceInProcessClient } from "./service-in-process-client";
 export { testStartupCheckHappyPath } from "./startup-check-tester";
 export * from "./generate-key-pairs";
+export { SimpleStorageContract, EthereumDriver } from "./ethereum-generator";
